// src/app/models/news-details.ts
// import { News } from './news';
export class NewsDetail {

    body: string;
    source_link: string;
    id: string;
    title: string;
    sub_heading: string;
    publication_date: string;
    picture: string;
    source_name: string;
    city: string;
    positive_votes: string;
    negative_votes: string;
    is_bookmarked: boolean;
    total_votes: string;
    positive_percentage: string;
    negative_percentage: string;
    negative_voted: boolean;
    positive_voted: boolean;
    summary: string;
    unsure_percentage: string;
    unsure_votes: string;
    unsure: boolean;
}